+++ src/PreservedTaintedness.php
@@ -11,6 +11,9 @@ class PreservedTaintedness {
 	/** @var int Combination of the class constants */
 	private $flags;
 
+	/** @var ParamLinksOffsets */
+	private $ownOffsets;
+
 	/** @var self[] Taintedness for each possible array element */
 	private $dimTaint = [];
 
@@ -24,9 +27,11 @@ class PreservedTaintedness {
 
 	/**
 	 * @param int $flags
+	 * @param ParamLinksOffsets|null $offsets
 	 */
-	public function __construct( int $flags ) {
+	public function __construct( int $flags, ParamLinksOffsets $offsets = null ) {
 		$this->flags = $flags;
+		$this->ownOffsets = $offsets ?? new ParamLinksOffsets();
 	}
 
 	/**
@@ -70,6 +75,7 @@ class PreservedTaintedness {
 	 */
 	public function mergeWith( self $other ) : void {
 		$this->flags |= $other->flags;
+		$this->ownOffsets->mergeWith( $other->ownOffsets );
 		$this->keysTaint |= $other->keysTaint;
 		if ( $other->unknownDimsTaint && !$this->unknownDimsTaint ) {
 			$this->unknownDimsTaint = $other->unknownDimsTaint;
@@ -96,13 +102,40 @@ class PreservedTaintedness {
 	}
 
 	/**
+	 * @param Taintedness $argTaint
+	 * @return Taintedness
+	 */
+	public function asTaintednessForArgument( Taintedness $argTaint ) : Taintedness {
+		if ( $this->flags & SecurityCheckPlugin::PRESERVE_TAINT ) {
+			$ret = $this->ownOffsets->appliedToTaintedness( $argTaint );
+		} elseif ( $this->flags ) {
+			$ret = $argTaint->withOnly( $this->flags );
+		} else {
+			$ret = new Taintedness( SecurityCheckPlugin::NO_TAINT );
+		}
+
+		foreach ( $this->dimTaint as $k => $val ) {
+			$ret->setOffsetTaintedness( $k, $val->asTaintednessForArgument( $argTaint ) );
+		}
+		if ( $this->unknownDimsTaint ) {
+			$ret->setOffsetTaintedness( null, $this->unknownDimsTaint->asTaintednessForArgument( $argTaint ) );
+		}
+		if ( $this->keysTaint & SecurityCheckPlugin::PRESERVE_TAINT ) {
+			$ret->addKeysTaintedness( $argTaint->get() );
+		} elseif ( $this->keysTaint ) {
+			$ret->addKeysTaintedness( $argTaint->get() & $this->keysTaint );
+		}
+		return $ret;
+	}
+
+	/**
 	 * Get a stringified representation of this taintedness suitable for the debug annotation
 	 *
 	 * @return string
 	 */
 	public function toShortString() : string {
 		$flags = SecurityCheckPlugin::taintToString( $this->flags );
-		$ret = "{Own: $flags";
+		$ret = "{Own: $flags, Offsets: " . $this->ownOffsets->__toString();
 		if ( $this->keysTaint ) {
 			$ret .= '; Keys: ' . SecurityCheckPlugin::taintToString( $this->keysTaint );
 		}
@@ -126,6 +159,7 @@ class PreservedTaintedness {
 	 * Make sure to clone member variables, too.
 	 */
 	public function __clone() {
+		$this->ownOffsets = clone $this->ownOffsets;
 		if ( $this->unknownDimsTaint ) {
 			$this->unknownDimsTaint = clone $this->unknownDimsTaint;
 		}