+++ tests/integration/msg2/test.php
@@ -10,7 +10,7 @@ class Message {
 	public function escaped() {
 		return 'placeholder';
 	}
-	public function rawParams( $params ) {
+	public function rawParams( ...$params ) {
 		return $this;
 	}
 	public function params( $params ) {